+++ src/neptune/handler.py
@@ -168,7 +168,7 @@ class Handler(SupportsNamespaces):
             >>> pretraining_run.stop()  # The root run is stopped
 
         For more information, see the docs:
-        https://docs.neptune.ai/api/field_types/#get_root_object
+        https://docs-legacy.neptune.ai/api/field_types/#get_root_object
         """
         return self._container
 
@@ -212,7 +212,7 @@ class Handler(SupportsNamespaces):
             >>> run['parameters'] = params
 
         For more information, see the docs:
-           https://docs.neptune.ai/api-reference/field-types
+           https://docs-legacy.neptune.ai/api-reference/field-types
         """
         with self._container.lock():
             attr = self._container.get_attribute(self._path)
@@ -254,7 +254,7 @@ class Handler(SupportsNamespaces):
             >>> run["dataset/data_sample"].upload(File("sample_data.csv"))
 
         For more information, see the docs:
-           https://docs.neptune.ai/api/field_types#upload
+           https://docs-legacy.neptune.ai/api/field_types#upload
 
         """
         value = FileVal.create_from(value, **kwargs)
@@ -312,7 +312,7 @@ class Handler(SupportsNamespaces):
                 Defaults to `False`.
 
         For more information, see the docs:
-           https://docs.neptune.ai/api-reference/field-types
+           https://docs-legacy.neptune.ai/api-reference/field-types
 
         """
         verify_type("step", step, (int, float, type(None)))
@@ -373,7 +373,7 @@ class Handler(SupportsNamespaces):
             * `FileSeries` - series of files
 
         When you log the first value, the type of the value determines what type of field is created.
-        To learn more about field types, see the docs: https://docs.neptune.ai/api/field_types
+        To learn more about field types, see the docs: https://docs-legacy.neptune.ai/api/field_types
 
         Args:
             value: Value to be added to the series field.
@@ -381,7 +381,7 @@ class Handler(SupportsNamespaces):
             timestamp: Optional time index of the log entry being appended, in Unix time format.
                 If None, the current time (obtained with `time.time()`) is used.
             wait: If True, the client sends all tracked metadata to the server before executing the call.
-                For more information, see: https://docs.neptune.ai/api/universal/#wait
+                For more information, see: https://docs-legacy.neptune.ai/api/universal/#wait
 
         Examples:
             >>> import neptune
@@ -422,7 +422,7 @@ class Handler(SupportsNamespaces):
             * `FileSeries` - series of files
 
         When you log the first value, the type of the value determines what type of field is created.
-        To learn more about field types, see the docs: https://docs.neptune.ai/api/field_types
+        To learn more about field types, see the docs: https://docs-legacy.neptune.ai/api/field_types
 
         Args:
             values: Values to be added to the series field, as a dictionary or collection.
@@ -430,7 +430,7 @@ class Handler(SupportsNamespaces):
             timestamps: Optional collection of time indeces for the entries being appended, in Unix time format.
                 If None, the current time (obtained with `time.time()`) is used.
             wait: If True, the client sends all tracked metadata to the server before executing the call.
-                For details, see https://docs.neptune.ai/api/universal/#wait
+                For details, see https://docs-legacy.neptune.ai/api/universal/#wait
 
         Example:
             The following example reads a CSV file into a pandas DataFrame and extracts the values
@@ -471,7 +471,7 @@ class Handler(SupportsNamespaces):
                 Defaults to `False`.
 
         For more information, see the docs:
-           https://docs.neptune.ai/api/field_types#add
+           https://docs-legacy.neptune.ai/api/field_types#add
         """
         verify_type("values", values, (str, Iterable))
         with self._container.lock():
@@ -495,7 +495,7 @@ class Handler(SupportsNamespaces):
             >>> run = neptune.init_run(with_id="RUN-100")
             >>> run["large_dataset"].pop()
 
-        See also the API reference: https://docs.neptune.ai/api/client_index/#pop
+        See also the API reference: https://docs-legacy.neptune.ai/api/client_index/#pop
         """
         with self._container.lock():
             handler = self
@@ -526,7 +526,7 @@ class Handler(SupportsNamespaces):
                 Defaults to `False`.
 
         For more information, see the docs:
-           https://docs.neptune.ai/api/field_types#remove
+           https://docs-legacy.neptune.ai/api/field_types#remove
         """
         return self._pass_call_to_attr(function_name="remove", values=values, wait=wait)
 
@@ -540,7 +540,7 @@ class Handler(SupportsNamespaces):
                 Defaults to `False`.
 
         For more information, see the docs:
-           https://docs.neptune.ai/api/field_types#clear
+           https://docs-legacy.neptune.ai/api/field_types#clear
         """
         return self._pass_call_to_attr(function_name="clear", wait=wait)
 
@@ -562,7 +562,7 @@ class Handler(SupportsNamespaces):
             If called on a namespace, returns a dictionary containing the values of all non-Atom fields.
 
         For more information on field types, see the docs:
-           https://docs.neptune.ai/api-reference/field-types
+           https://docs-legacy.neptune.ai/api-reference/field-types
         """
         return self._pass_call_to_attr(function_name="fetch")
 
@@ -578,7 +578,7 @@ class Handler(SupportsNamespaces):
             Fetches the last value stored in the series from Neptune.
 
         For more information on field types, see the docs:
-           https://docs.neptune.ai/api-reference/field-types
+           https://docs-legacy.neptune.ai/api-reference/field-types
         """
         return self._pass_call_to_attr(function_name="fetch_last")
 
@@ -608,7 +608,7 @@ class Handler(SupportsNamespaces):
             ``Pandas.DataFrame``: containing all the values and their indexes stored in the series field.
 
         For more information on field types, see the docs:
-           https://docs.neptune.ai/api-reference/field-types
+           https://docs-legacy.neptune.ai/api-reference/field-types
         """
         return self._pass_call_to_attr(
             function_name="fetch_values",
@@ -630,7 +630,7 @@ class Handler(SupportsNamespaces):
                 Defaults to `None`.
 
         For more information, see the docs:
-            https://docs.neptune.ai/api/field_types#delete_files
+            https://docs-legacy.neptune.ai/api/field_types#delete_files
         """
         return self._pass_call_to_attr(function_name="delete_files", paths=paths, wait=wait)
 
@@ -663,7 +663,7 @@ class Handler(SupportsNamespaces):
                 Defaults to `None`.
 
         For more information, see the docs:
-           https://docs.neptune.ai/api-reference/field-types
+           https://docs-legacy.neptune.ai/api-reference/field-types
         """
         return self._pass_call_to_attr(function_name="download", destination=destination, progress_bar=progress_bar)
 
@@ -679,7 +679,7 @@ class Handler(SupportsNamespaces):
                 Defaults to `None`.
 
         For more information, see the docs:
-           https://docs.neptune.ai/api/field_types#download_last
+           https://docs-legacy.neptune.ai/api/field_types#download_last
         """
         return self._pass_call_to_attr(function_name="download_last", destination=destination)
 
@@ -687,7 +687,7 @@ class Handler(SupportsNamespaces):
         """Fetches the hash of an artifact.
 
         You may also want to check the docs:
-           https://docs.neptune.ai/api/field_types#fetch_hash
+           https://docs-legacy.neptune.ai/api/field_types#fetch_hash
         """
         return self._pass_call_to_attr(function_name="fetch_hash")
 
@@ -695,7 +695,7 @@ class Handler(SupportsNamespaces):
         """Fetches the extension of a file.
 
         You may also want to check the docs:
-           https://docs.neptune.ai/api/field_types#fetch_extension
+           https://docs-legacy.neptune.ai/api/field_types#fetch_extension
         """
         return self._pass_call_to_attr(function_name="fetch_extension")
 
@@ -703,7 +703,7 @@ class Handler(SupportsNamespaces):
         """Fetches the list of files in an artifact and their metadata.
 
         You may also want to check the docs:
-           https://docs.neptune.ai/api/field_types#fetch_files_list
+           https://docs-legacy.neptune.ai/api/field_types#fetch_files_list
         """
         return self._pass_call_to_attr(function_name="fetch_files_list")
 
@@ -740,7 +740,7 @@ class Handler(SupportsNamespaces):
             8, 17, 10, 31, 26, 338000, tzinfo=tzutc()), file_type='file'), ...]
 
         For more information, see the API reference:
-           https://docs.neptune.ai/api/field_types#list_fileset_files
+           https://docs-legacy.neptune.ai/api/field_types#list_fileset_files
         """
         return self._pass_call_to_attr(function_name="list_fileset_files", path=path)
 
@@ -752,7 +752,7 @@ class Handler(SupportsNamespaces):
         """Creates an artifact tracking some files.
 
         You may also want to check the docs:
-           https://docs.neptune.ai/api/field_types#track_files
+           https://docs-legacy.neptune.ai/api/field_types#track_files
         """
         with self._container.lock():
             attr = self._container.get_attribute(self._path)