+++ src/neptune/metadata_containers/run.py
@@ -264,19 +264,19 @@ class Run(MetadataContainer):
 
     For more, see the docs:
         Initializing a run:
-            https://docs.neptune.ai/api/neptune#init_run
+            https://docs-legacy.neptune.ai/api/neptune#init_run
         Run class reference:
-            https://docs.neptune.ai/api/run/
+            https://docs-legacy.neptune.ai/api/run/
         Essential logging methods:
-            https://docs.neptune.ai/logging/methods/
+            https://docs-legacy.neptune.ai/logging/methods/
         Resuming a run:
-            https://docs.neptune.ai/logging/to_existing_object/
+            https://docs-legacy.neptune.ai/logging/to_existing_object/
         Setting a custom run ID:
-            https://docs.neptune.ai/logging/custom_run_id/
+            https://docs-legacy.neptune.ai/logging/custom_run_id/
         Logging to multiple runs at once:
-            https://docs.neptune.ai/logging/to_multiple_objects/
+            https://docs-legacy.neptune.ai/logging/to_multiple_objects/
         Accessing the run from multiple places:
-            https://docs.neptune.ai/logging/from_multiple_places/
+            https://docs-legacy.neptune.ai/logging/from_multiple_places/
     """
 
     container_type = ContainerType.RUN
@@ -570,7 +570,7 @@ def capture_only_if_non_interactive(mode) -> bool:
                 " You can set them to 'True' when initializing the run and the monitoring will"
                 " continue until you call run.stop() or the kernel stops."
                 " NOTE: To track the source files, pass their paths to the 'source_code'"
-                " argument. For help, see: https://docs.neptune.ai/logging/source_code/",
+                " argument. For help, see: https://docs-legacy.neptune.ai/logging/source_code/",
                 exception=NeptuneWarning,
             )
         return False